header hiding on scroll, moved css, updated docs
