+++ src/components/Site.js
@@ -1,14 +1,17 @@
 import React,{Component} from 'react'
 import {connect} from 'react-redux'
 import {bindActionCreators} from 'redux'
+
 import { withRouter } from 'react-router-dom'
-import Navbar from './navbar/Navbar'
 import {checkLogin, logout, userSyncSuccess} from '../actions/auth-actions'
 import {login} from '../lib/auth'
+
 import SyncUser from './SyncUser'
-import {Grid} from 'react-flexbox-grid'
 import BottomBar from './BottomBar'
-import '../css/layout.css'
+import Navbar from './navbar/Navbar'
+import {Grid} from 'react-flexbox-grid'
+import Headroom from 'react-headroom'
+import '../styles/css/layout.css'
 
 class Site extends Component {
 
@@ -54,11 +57,13 @@ class Site extends Component {
     return(
       <div style={this.styles.main}>
         {this.renderSyncUser()}
+      <Headroom style={{zIndex:2}}>
         <Navbar handleLogout={logout}
           handleLogin={login}
           isAuthenticated={isAuthenticated}
           profile={profile}
         />
+      </Headroom>
         <Grid style={{marginTop:100}}>
           {children}
         </Grid>
