+++ src/components/navbar/Navbar.js
@@ -19,8 +19,7 @@ const Navbar = (props) => {
     },
     navbar: {
       backgroundColor:"#ffffff",
-      position: "fixed",
-      top: 0,
+      zIndex: 2
     },
     accountCircleIcon: {
       width: 30,
@@ -55,12 +54,14 @@ const Navbar = (props) => {
   const userControls = renderUserControls()
 
   return(
+  <div style={{zIndex:2}}>
     <AppBar
       title={<span style={styles.title}>KOVI</span>}
       onTitleTouchTap={handleTouchTap}
       style={styles.navbar}
       iconElementRight={userControls}
     />
+  </div>
   )
 }
